Fix gcc compiler warnings from -Wall

+++ src/Battlescape/BattlescapeGenerator.cpp
@@ -1620,7 +1620,7 @@ int BattlescapeGenerator::loadMAP(MapBlock *mapblock, int xoff, int yoff, RuleTe
 	unsigned char value[4];
 	std::ostringstream filename;
 	filename << "MAPS/" << mapblock->getName() << ".MAP";
-	int terrainObjectID;
+	unsigned int terrainObjectID;
 
 	// Load file
 	std::ifstream mapFile (CrossPlatform::getDataFile(filename.str()).c_str(), std::ios::in| std::ios::binary);
@@ -1666,11 +1666,11 @@ int BattlescapeGenerator::loadMAP(MapBlock *mapblock, int xoff, int yoff, RuleTe
 	{
 		for (int part = 0; part < 4; part++)
 		{
-			terrainObjectID = (int)((unsigned char)value[part]);
+			terrainObjectID = ((unsigned char)value[part]);
 			if (terrainObjectID>0)
 			{
 				int mapDataSetID = mapDataSetOffset;
-				int mapDataID = terrainObjectID;
+				unsigned int mapDataID = terrainObjectID;
 				MapData *md = terrain->getMapData(&mapDataID, &mapDataSetID);
 				_save->getTile(Position(x, y, z))->setMapData(md, mapDataID, mapDataSetID, part);
 			}
